+++ package.json
@@ -13,8 +13,10 @@
     "@mantine/core": "^8.0.2",
     "@mantine/hooks": "^8.0.2",
     "axios": "^1.9.0",
+    "jwt-decode": "^4.0.0",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
